% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warabandi_data.R
\docType{data}
\name{warabandi_data}
\alias{warabandi_data}
\title{Warabandi data required for calculation}
\format{
It comprises name of farmer, Rakva, bharai, katai etc
}
\usage{
data(warabandi_data)
}
\description{
Warabandi data required for calculation
}
\keyword{datasets}
